requirements.txt
python-telegram-bot==21.7
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
openai==1.58.0
aiohttp==3.11.0
pgvector==0.3.6
jinja2==3.1.4
python-multipart==0.0.17
pydantic==2.10.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.0
itsdangerous==2.2.0
apscheduler==3.10.4
